# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": ["code", "message", "requestId"],
        "type": "object"
      },
...
      "CardDraftListResponseDto": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/CardDraftViewDto"
            },
            "type": "array"
          },
          "nextCursor": {
            "format": "uuid",
            "nullable": true,
            "type": "object"
          }
        },
        "required": ["data"],
        "type": "object"
      },
      "CardDraftResponseDto": {
        "properties": {
          "cardDraft": {
            "$ref": "#/components/schemas/CardDraftViewDto"
          }
        },
        "required": ["cardDraft"],
        "type": "object"
      },
      "CardDraftViewDto": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "creationRoute": {
            "enum": ["personalize_template", "build_my_card"],
            "type": "string"
          },
          "creativeBrief": {
            "additionalProperties": true,
            "type": "object"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "occasion": {
            "nullable": true,
            "type": "object"
          },
          "relationship": {
            "nullable": true,
            "type": "object"
          },
          "revisionNumber": {
            "minimum": 1,
            "type": "integer"
          },
          "status": {
            "enum": ["draft", "generating", "review", "approved", "ordered", "sent", "archived"],
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": ["id", "creationRoute", "status", "creativeBrief", "revisionNumber", "createdAt", "updatedAt"],
        "type": "object"
      },
...
      "CreateCardDraftDto": {
        "properties": {
          "creationRoute": {
            "enum": ["personalize_template", "build_my_card"],
            "type": "string"
          },
          "creativeBrief": {
            "additionalProperties": true,
            "type": "object"
          },
          "occasion": {
            "maxLength": 160,
            "type": "string"
          },
          "relationship": {
            "maxLength": 160,
            "type": "string"
          }
        },
        "required": ["creationRoute"],
        "type": "object"
      },
...
      "PublicCardResponseDto": {
        "properties": {
          "card": {
            "$ref": "#/components/schemas/PublicCardViewDto"
          }
        },
        "required": ["card"],
        "type": "object"
      },
      "PublicCardViewDto": {
        "properties": {
          "expiresAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "publicPath": {
            "type": "string"
          },
          "qrPayloadVersion": {
            "minimum": 1,
            "type": "integer"
          },
          "song": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PublicSongViewDto"
              }
            ],
            "nullable": true,
            "type": "object"
          }
        },
        "required": ["id", "publicPath", "qrPayloadVersion"],
        "type": "object"
      },
      "PublicSongViewDto": {
        "properties": {
          "durationSeconds": {
            "minimum": 0,
            "nullable": true,
            "type": "number"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": ["pending", "generating", "ready", "failed"],
            "type": "string"
          }
        },
        "required": ["id", "status"],
        "type": "object"
      },
...
      "UpdateCardDraftDto": {
        "properties": {
          "creativeBrief": {
            "additionalProperties": true,
            "type": "object"
          },
          "occasion": {
            "maxLength": 160,
            "type": "string"
          },
          "relationship": {
            "maxLength": 160,
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "Cognito access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Authenticated Canada-first Souvenote MVP API",
    "title": "Souvenote API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/card-drafts": {
      "get": {
        "operationId": "listCardDrafts",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "Opaque owner-scoped cursor from the previous page.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardDraftListResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request conflict"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": ["card-drafts"]
      },
      "post": {
        "operationId": "createCardDraft",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCardDraftDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardDraftResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request conflict"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": ["card-drafts"]
      }
    },
    "/api/v1/card-drafts/{draftId}": {
      "get": {
        "operationId": "getCardDraft",
        "parameters": [
          {
            "in": "path",
            "name": "draftId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardDraftResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request conflict"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": ["card-drafts"]
      },
      "patch": {
        "operationId": "updateCardDraft",
        "parameters": [
          {
            "in": "path",
            "name": "draftId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCardDraftDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CardDraftResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request conflict"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": ["card-drafts"]
      }
    },
...
    "/api/v1/public/cards/{shareToken}": {
      "get": {
        "operationId": "getPublicCard",
        "parameters": [
          {
            "in": "path",
            "name": "shareToken",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 32,
              "pattern": "^[A-Za-z0-9_-]+$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicCardResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request conflict"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "summary": "",
        "tags": ["public-cards"]
      }
    },
...
  },
  "servers": [],
...
}